+++ realm/realm-library/src/main/java/io/realm/RealmList.java
@@ -50,16 +50,18 @@ import rx.Observable;
  * @param <E> the class of objects in list.
  */
 
-public class RealmList<E extends RealmModel> extends AbstractList<E> implements OrderedRealmCollection<E> {
+public class RealmList<E extends RealmModel> extends AbstractList<E>
+        implements OrderedRealmCollection<E>,
+        RealmCollectionObservable<RealmList<E>, OrderedRealmCollectionChangeListener<RealmList<E>>> {
 
     private static final String ONLY_IN_MANAGED_MODE_MESSAGE = "This method is only available in managed mode";
     private static final String NULL_OBJECTS_NOT_ALLOWED_MESSAGE = "RealmList does not accept null values";
     public static final String REMOVE_OUTSIDE_TRANSACTION_ERROR = "Objects can only be removed from inside a write transaction";
 
-    private final boolean managedMode;
+    private final io.realm.internal.Collection collection;
     protected Class<E> clazz;
     protected String className;
-    protected LinkView view;
+    final LinkView view;
     protected BaseRealm realm;
     private List<E> unmanagedList;
 
@@ -71,7 +73,8 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      * Use {@link io.realm.Realm#copyToRealm(Iterable)} to properly persist its elements in Realm.
      */
     public RealmList() {
-        managedMode = false;
+        collection = null;
+        view = null;
         unmanagedList = new ArrayList<E>();
     }
 
@@ -88,7 +91,8 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
         if (objects == null) {
             throw new IllegalArgumentException("The objects argument cannot be null");
         }
-        managedMode = false;
+        collection = null;
+        view = null;
         unmanagedList = new ArrayList<E>(objects.length);
         Collections.addAll(unmanagedList, objects);
     }
@@ -101,14 +105,14 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      * @param realm reference to Realm containing the data.
      */
     RealmList(Class<E> clazz, LinkView linkView, BaseRealm realm) {
-        this.managedMode = true;
+        this.collection = new io.realm.internal.Collection(realm.sharedRealm, linkView, null);
         this.clazz = clazz;
         this.view = linkView;
         this.realm = realm;
     }
 
     RealmList(String className, LinkView linkView, BaseRealm realm) {
-        this.managedMode = true;
+        this.collection = new io.realm.internal.Collection(realm.sharedRealm, linkView, null);
         this.view = linkView;
         this.realm = realm;
         this.className = className;
@@ -161,7 +165,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     @Override
     public void add(int location, E object) {
         checkValidObject(object);
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             if (location < 0 || location > size()) {
                 throw new IndexOutOfBoundsException("Invalid index " + location + ", size is " + size());
@@ -193,7 +197,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     @Override
     public boolean add(E object) {
         checkValidObject(object);
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             RealmObjectProxy proxy = (RealmObjectProxy) copyToRealmIfNeeded(object);
             view.add(proxy.realmGet$proxyState().getRow$realm().getIndex());
@@ -226,7 +230,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     public E set(int location, E object) {
         checkValidObject(object);
         E oldObject;
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             RealmObjectProxy proxy = (RealmObjectProxy) copyToRealmIfNeeded(object);
             oldObject = get(location);
@@ -294,7 +298,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      * @throws java.lang.IndexOutOfBoundsException if any position is outside [0, size()].
      */
     public void move(int oldPos, int newPos) {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             view.move(oldPos, newPos);
         } else {
@@ -319,7 +323,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public void clear() {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             view.clear();
         } else {
@@ -339,7 +343,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     @Override
     public E remove(int location) {
         E removedItem;
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             removedItem = get(location);
             view.remove(location);
@@ -369,7 +373,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean remove(Object object) {
-        if (managedMode && !realm.isInTransaction()) {
+        if (isManaged() && !realm.isInTransaction()) {
             throw new IllegalStateException(REMOVE_OUTSIDE_TRANSACTION_ERROR);
         }
         return super.remove(object);
@@ -393,7 +397,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean removeAll(Collection<?> collection) {
-        if (managedMode && !realm.isInTransaction()) {
+        if (isManaged() && !realm.isInTransaction()) {
             throw new IllegalStateException(REMOVE_OUTSIDE_TRANSACTION_ERROR);
         }
         return super.removeAll(collection);
@@ -404,7 +408,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean deleteFirstFromRealm() {
-        if (managedMode) {
+        if (isManaged()) {
             if (size() > 0) {
                 deleteFromRealm(0);
                 modCount++;
@@ -422,7 +426,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean deleteLastFromRealm() {
-        if (managedMode) {
+        if (isManaged()) {
             if (size() > 0) {
                 deleteFromRealm(size() - 1);
                 modCount++;
@@ -445,7 +449,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public E get(int location) {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             long rowIndex = view.getTargetRowIndex(location);
             return realm.get(clazz, className, rowIndex);
@@ -469,7 +473,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     }
 
     private E firstImpl(boolean shouldThrow, E defaultValue) {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             if (!view.isEmpty()) {
                 return get(0);
@@ -500,7 +504,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     }
 
     private E lastImpl(boolean shouldThrow, E defaultValue) {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             if (!view.isEmpty()) {
                 return get((int) view.size() - 1);
@@ -529,7 +533,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public RealmResults<E> sort(String fieldName, Sort sortOrder) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().findAllSorted(fieldName, sortOrder);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -549,7 +553,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public RealmResults<E> sort(String[] fieldNames, Sort[] sortOrders) {
-        if (managedMode) {
+        if (isManaged()) {
             return where().findAllSorted(fieldNames, sortOrders);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -561,7 +565,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public void deleteFromRealm(int location) {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             view.removeTargetRow(location);
             modCount++;
@@ -578,7 +582,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public int size() {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             long size = view.size();
             return size < Integer.MAX_VALUE ? (int) size : Integer.MAX_VALUE;
@@ -595,7 +599,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      * @see io.realm.RealmQuery
      */
     public RealmQuery<E> where() {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             return RealmQuery.createQueryFromList(this);
         } else {
@@ -608,7 +612,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Number min(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().min(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -620,7 +624,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Number max(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().max(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -632,7 +636,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Number sum(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().sum(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -644,7 +648,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public double average(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().average(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -656,7 +660,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Date maxDate(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().maximumDate(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -668,7 +672,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Date minDate(String fieldName) {
-        if (managedMode) {
+        if (isManaged()) {
             return this.where().minimumDate(fieldName);
         } else {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
@@ -680,7 +684,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean deleteAllFromRealm() {
-        if (managedMode) {
+        if (isManaged()) {
             checkValidView();
             if (size() > 0) {
                 view.removeAllTargetRows();
@@ -722,7 +726,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public boolean contains(Object object) {
-        if (managedMode) {
+        if (isManaged()) {
             realm.checkIfValid();
 
             // Deleted objects can never be part of a RealmList
@@ -749,7 +753,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public Iterator<E> iterator() {
-        if (managedMode) {
+        if (isManaged()) {
             return new RealmItr();
         } else {
             return super.iterator();
@@ -769,7 +773,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public ListIterator<E> listIterator(int location) {
-        if (managedMode) {
+        if (isManaged()) {
             return new RealmListItr(location);
         } else {
             return super.listIterator(location);
@@ -801,7 +805,7 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
      */
     @Override
     public OrderedRealmCollectionSnapshot<E> createSnapshot() {
-        if (!managedMode) {
+        if (!isManaged()) {
             throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
         }
         checkValidView();
@@ -819,13 +823,13 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(managedMode ? clazz.getSimpleName() : getClass().getSimpleName());
+        sb.append(isManaged() ? clazz.getSimpleName() : getClass().getSimpleName());
         sb.append("@[");
-        if (managedMode && !isAttached()) {
+        if (isManaged() && !isAttached()) {
             sb.append("invalid");
         } else {
             for (int i = 0; i < size(); i++) {
-                if (managedMode) {
+                if (isManaged()) {
                     sb.append(((RealmObjectProxy) get(i)).realmGet$proxyState().getRow$realm().getIndex());
                 } else {
                     sb.append(System.identityHashCode(get(i)));
@@ -878,6 +882,59 @@ public class RealmList<E extends RealmModel> extends AbstractList<E> implements 
         } else {
             throw new UnsupportedOperationException(realm.getClass() + " does not support RxJava.");
         }
+    }
+
+    private void checkForAddRemoveListener(Object listener, boolean checkListener) {
+        if (checkListener && listener == null) {
+            throw new IllegalArgumentException("Listener should not be null");
+        }
+        realm.checkIfValid();
+        realm.sharedRealm.capabilities.checkCanDeliverNotification(BaseRealm.LISTENER_NOT_ALLOWED_MESSAGE);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void addChangeListener(OrderedRealmCollectionChangeListener<RealmList<E>> listener) {
+        checkForAddRemoveListener(listener, true);
+        collection.addListener(this, listener);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void removeChangeListener(OrderedRealmCollectionChangeListener<RealmList<E>> listener) {
+        checkForAddRemoveListener(listener, true);
+        collection.removeListener(this, listener);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void addChangeListener(RealmChangeListener<RealmList<E>> listener) {
+        checkForAddRemoveListener(listener, true);
+        collection.addListener(this, listener);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void removeChangeListener(RealmChangeListener<RealmList<E>> listener) {
+        checkForAddRemoveListener(listener, true);
+        collection.removeListener(this, listener);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void removeAllChangeListeners() {
+        checkForAddRemoveListener(null, false);
+        collection.removeAllListeners();
     }
 
     // Custom RealmList iterator.