+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 
 * Added support for sorting by link's field (#672).
 * Added `OrderedRealmCollectionSnapshot` class and `OrderedRealmCollection.createSnapshot()` method. `OrderedRealmCollectionSnapshot` is useful when changing `RealmResults` or `RealmList` in simple loops.
-* Added `RealmList.asObservable()`
+* Added support for adding listeners on `RealmList`.
+* Added `RealmList.asObservable()`.
 
 ### Internal